+++ crates/algod_client/src/models/debug_settings_prof.rs
@@ -16,10 +16,10 @@ use serde::{Deserialize, Serialize};
 pub struct DebugSettingsProf {
     /// The rate of blocking events. The profiler aims to sample an average of one blocking event per rate nanoseconds spent blocked. To turn off profiling entirely, pass rate 0.
     #[serde(rename = "block-rate", skip_serializing_if = "Option::is_none")]
-    pub block_rate: Option<i32>,
+    pub block_rate: Option<u64>,
     /// The rate of mutex events. On average 1/rate events are reported. To turn off profiling entirely, pass rate 0
     #[serde(rename = "mutex-rate", skip_serializing_if = "Option::is_none")]
-    pub mutex_rate: Option<i32>,
+    pub mutex_rate: Option<u64>,
 }
 
 impl DebugSettingsProf {
