+++ crates/algod_client/src/models/simulation_opcode_trace_unit.rs
@@ -21,7 +21,7 @@ use crate::models::ScratchChange;
 pub struct SimulationOpcodeTraceUnit {
     /// The program counter of the current opcode being evaluated.
     #[serde(rename = "pc")]
-    pub pc: i32,
+    pub pc: u64,
     /// The writes into scratch slots.
     #[serde(rename = "scratch-changes", skip_serializing_if = "Option::is_none")]
     pub scratch_changes: Option<Vec<ScratchChange>>,
@@ -30,10 +30,10 @@ pub struct SimulationOpcodeTraceUnit {
     pub state_changes: Option<Vec<ApplicationStateOperation>>,
     /// The indexes of the traces for inner transactions spawned by this opcode, if any.
     #[serde(rename = "spawned-inners", skip_serializing_if = "Option::is_none")]
-    pub spawned_inners: Option<Vec<i32>>,
+    pub spawned_inners: Option<Vec<u64>>,
     /// The number of deleted stack values by this opcode.
     #[serde(rename = "stack-pop-count", skip_serializing_if = "Option::is_none")]
-    pub stack_pop_count: Option<i32>,
+    pub stack_pop_count: Option<u64>,
     /// The values added by this opcode to the stack.
     #[serde(rename = "stack-additions", skip_serializing_if = "Option::is_none")]
     pub stack_additions: Option<Vec<AvmValue>>,
@@ -45,7 +45,7 @@ impl AlgorandMsgpack for SimulationOpcodeTraceUnit {
 
 impl SimulationOpcodeTraceUnit {
     /// Constructor for SimulationOpcodeTraceUnit
-    pub fn new(pc: i32) -> SimulationOpcodeTraceUnit {
+    pub fn new(pc: u64) -> SimulationOpcodeTraceUnit {
         SimulationOpcodeTraceUnit {
             pc,
             scratch_changes: None,
